+++ src/triage/component/catwalk/individual_importance/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-from triage.component.catwalk.utils import save_db_objects
+from triage.util.db import save_db_objects
 from triage.component.results_schema import IndividualImportance
 
 from .uniform import uniform_distribution
