+++ src/triage/component/catwalk/ranking.py
