+++ src/triage/experiments/multicore.py
@@ -4,7 +4,7 @@ from functools import partial
 from pebble import ProcessPool
 from multiprocessing.reduction import ForkingPickler
 
-from triage.component.catwalk.utils import Batch
+from triage.util.iteration import Batch
 
 from triage.experiments import ExperimentBase
